requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pymupdf==1.22.1
pytesseract==0.3.10
Pillow==9.5.0
spacy==3.7.1
python-multipart==0.0.6
aiofiles==23.1.0
cryptography==41.0.3
